Add Desired Delivery Month to Orders Add new Metallic Paint discount to Orders if model has it listed Add commands to clean up Orders, Vehicles and Customers who have been listed as deleted for longer than a month

+++ routes/console.php
@@ -1,8 +1,10 @@
 <?php
 
+use App\Models\Customer;
 use App\Models\Delivery;
 use App\Models\Reservation;
 use App\Models\Vehicle;
+use App\Models\Order;
 use Carbon\Carbon;
 use Illuminate\Foundation\Inspiring;
 
@@ -22,14 +24,69 @@ Artisan::command('inspire', function () {
 })->describe('Display an inspiring quote');
 
 Artisan::command('cleanVehicle', function () {
-    $date = Carbon::now()->subDays(30);
-    $deleted = Vehicle::where('deleted_at', '>', $date)
+    $date = Carbon::now()->subMonth();
+    $deleted = Vehicle::where('deleted_at', '<', $date)
         ->onlyTrashed()
         ->get();
 
-    foreach ($deleted as $item) {
-        $item->forceDelete();
+    if (count($deleted) !== 0) {
+        $this->line('Clearing vehicles deleted more than a month ago...');
+        $bar = $this->output->createProgressBar(count($deleted));
+        $bar->start();
+        foreach ($deleted as $item) {
+            $item->forceDelete();
+            $bar->advance();
+        }
+        $bar->finish();
+    } else {
+        $this->line('No Vehicles require cleaning');
     }
+})->purpose(
+    'Remove Vehicles if they are still in the recycle bin after a month',
+);
+
+Artisan::command('cleanOrders', function () {
+    $date = Carbon::now()->subMonth();
+    $deleted = Order::where('deleted_at', '<', $date)
+        ->onlyTrashed()
+        ->get();
+    if (count($deleted) !== 0) {
+        $this->line('Clearing orders deleted more than a month ago...');
+        $bar = $this->output->createProgressBar(count($deleted));
+        $bar->start();
+        foreach ($deleted as $item) {
+            $item->forceDelete();
+            $bar->advance();
+        }
+        $bar->finish();
+    } else {
+        $this->line('No Orders require cleaning');
+    }
+})->purpose('Remove orders if they are still in the recycle bin after a month');
+
+Artisan::command('cleanCustomers', function () {
+    $date = Carbon::now()->subMonth();
+    $deleted = Customer::where('deleted_at', '<', $date)
+        ->onlyTrashed()
+        ->get();
+    if (count($deleted) !== 0) {
+        $this->line('Clearing customers deleted more than a month ago...');
+        $bar = $this->output->createProgressBar(count($deleted));
+        $bar->start();
+        foreach ($deleted as $item) {
+            $item->forceDelete();
+            $bar->advance();
+        }
+        $bar->finish();
+    } else {
+        $this->line('No Customers require cleaning');
+    }
+})->purpose('Remove orders if they are still in the recycle bin after a month');
+
+Artisan::command('databaseClean', function () {
+    $this->call('cleanOrders');
+    $this->call('cleanVehicle');
+    $this->call('cleanCustomers');
 });
 
 Artisan::command('checkReservationExpiry', function () {
